parseFloat añadido

+++ ex8.js
@@ -3,19 +3,19 @@ function operacions(op) {
 
     const ops = {
         sumar: function sumarNumeros(n1, n2) {
-            return (parseInt(n1) + parseInt(n2));
+            return (parseFloat(n1) + parseFloat(n2));
         },
 
         restar: function restarNumeros(n1, n2) {
-            return (parseInt(n1) - parseInt(n2));
+            return (parseFloat(n1) - parseFloat(n2));
         },
 
         multiplicar: function multiplicarNumeros(n1, n2) {
-            return (parseInt(n1) * parseInt(n2));
+            return (parseFloat(n1) * parseFloat(n2));
         },
 
         dividir: function dividirNumeros(n1, n2) {
-            return (parseInt(n1) / parseInt(n2));
+            return (parseFloat(n1) / parseFloat(n2));
         }
 
     };
